day_12/task_17/requirements.txt
streamlit
autogen
google-generativeai
python-dotenv
pandas
plotly-express